fix(gateway): reject unsafe audit scope transitions
