+++ crates/astrid-gateway/src/routes/audit.rs
@@ -40,6 +40,7 @@ const DEFAULT_LIMIT: usize = 100;
 const MAX_LIMIT: usize = 1000;
 const CURSOR_SCOPE_ALL: &str = "all";
 const CURSOR_SCOPE_PRINCIPAL_PREFIX: char = 'p';
+const CURSOR_SCOPE_CHANGED: &str = "audit visibility widened or changed principal during pagination; restart pagination without a cursor";
 
 /// Query parameters for `GET /api/sys/audit`. All optional —
 /// the default behaviour is "the last [`DEFAULT_LIMIT`] entries
@@ -278,7 +279,7 @@ pub async fn get_audit(
     // still accepted for compatibility.
     let cursor = parse_cursor(query.cursor.as_deref())?;
     validate_cursor_scope(cursor.2.as_ref(), &access)?;
-    let (page, next_cursor) = paginate_page(entries, &query, &access, cursor, limit);
+    let (page, next_cursor) = paginate_page(entries, &query, &access, cursor, limit)?;
 
     Ok(Json(AuditQueryResponse {
         entries: page,
@@ -323,19 +324,25 @@ fn validate_cursor_scope(
         return Ok(());
     };
     let current_scope = access.current_cursor_scope();
+    validate_scope_continuation(previous_scope, &current_scope)
+}
+
+fn validate_scope_continuation(
+    previous_scope: &CursorScope,
+    current_scope: &CursorScope,
+) -> GatewayResult<()> {
     if current_scope.accepts_continuation_from(previous_scope) {
         Ok(())
     } else {
-        Err(GatewayError::BadRequest(
-            "cursor scope widened or changed principal since the previous page; restart pagination without a cursor".into(),
-        ))
+        Err(GatewayError::BadRequest(CURSOR_SCOPE_CHANGED.into()))
     }
 }
 
 /// Walk the entries (newest first) and assemble one page worth of
 /// rendered views, honouring every stable filter plus the live
 /// principal restriction. Returns the page plus the next-page cursor
-/// (or `None` if the result is the last page). Pulled out of
+/// (or `None` if the result is the last page), and rejects a live
+/// scope widening that would invalidate the raw cursor boundary. Pulled out of
 /// [`get_audit`] so the handler stays inside the function-length
 /// budget and the cursor arithmetic lives in one place.
 fn paginate_page(
@@ -344,16 +351,18 @@ fn paginate_page(
     access: &AuditAccess<'_>,
     cursor: (Option<u64>, usize, Option<CursorScope>),
     limit: usize,
-) -> (Vec<AuditEntryView>, Option<String>) {
-    let (cursor_ts, cursor_offset, _) = cursor;
+) -> GatewayResult<(Vec<AuditEntryView>, Option<String>)> {
+    let (cursor_ts, cursor_offset, cursor_scope) = cursor;
     let mut page: Vec<AuditEntryView> = Vec::with_capacity(limit);
     let mut raw_last_ts: Option<u64> = None;
     let mut raw_ts_position: usize = 0;
     let mut last_visible_ts: Option<u64> = None;
     let mut last_visible_ts_position: usize = 0;
     let mut last_visible_scope: Option<CursorScope> = None;
-    let mut effective_scope: Option<CursorScope> = None;
-    let mut clipped_by_scope_change = false;
+    let mut effective_scope = access.current_cursor_scope();
+    if let Some(cursor_scope) = cursor_scope.as_ref() {
+        validate_scope_continuation(cursor_scope, &effective_scope)?;
+    }
     for entry in entries {
         let Some(view) = render_entry(&entry) else {
             continue;
@@ -381,6 +390,10 @@ fn paginate_page(
         };
         raw_last_ts = Some(view.ts_epoch);
 
+        let current_scope = access.current_cursor_scope();
+        validate_scope_continuation(&effective_scope, &current_scope)?;
+        effective_scope = current_scope;
+
         // Cursor positioning: drop everything strictly newer than
         // the cursor's `ts`, then skip the first `cursor_offset`
         // same-second entries in the stable pre-authorization order.
@@ -393,17 +406,7 @@ fn paginate_page(
             }
         }
 
-        let current_scope = access.current_cursor_scope();
-        if let Some(previous_scope) = effective_scope.as_ref()
-            && !current_scope.accepts_continuation_from(previous_scope)
-        {
-            clipped_by_scope_change = true;
-            break;
-        }
-        effective_scope = Some(current_scope);
-        if let Some(p) = effective_scope
-            .as_ref()
-            .and_then(CursorScope::principal_filter)
+        if let Some(p) = effective_scope.principal_filter()
             && view.principal.as_deref() != Some(p.as_str())
         {
             continue;
@@ -411,14 +414,14 @@ fn paginate_page(
 
         last_visible_ts = Some(view.ts_epoch);
         last_visible_ts_position = raw_ts_position;
-        last_visible_scope = effective_scope.clone();
+        last_visible_scope = Some(effective_scope.clone());
         page.push(view);
         if page.len() >= limit {
             break;
         }
     }
 
-    let next_cursor = if page.len() == limit || clipped_by_scope_change {
+    let next_cursor = if page.len() == limit {
         last_visible_ts
             .zip(last_visible_scope)
             .map(|(t, scope)| format!("{t}_{last_visible_ts_position}_{}", scope.encode()))
@@ -426,7 +429,7 @@ fn paginate_page(
         None
     };
 
-    (page, next_cursor)
+    Ok((page, next_cursor))
 }
 
 /// Parse an opaque cursor into `(ts_epoch, equal_ts_offset, scope)`.
@@ -492,427 +495,4 @@ fn render_entry(entry: &AuditEntry) -> Option<AuditEntryView> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    use std::sync::atomic::{AtomicUsize, Ordering};
-
-    use astrid_audit::{AuditLog, AuthorizationProof};
-    use astrid_core::{SessionId, Timestamp};
-    use astrid_crypto::KeyPair;
-    use chrono::TimeZone;
-
-    fn admin_action(method: &str, target: Option<&str>) -> AuditAction {
-        AuditAction::AdminRequest {
-            method: method.into(),
-            required_capability: "*".into(),
-            target_principal: target.map(|s| PrincipalId::new(s).unwrap()),
-            params: None,
-            device_key_id: None,
-        }
-    }
-
-    #[tokio::test]
-    async fn render_drops_non_admin_actions() {
-        // Non-admin entries (MCP tool calls, capsule events) belong
-        // to a different audit feed; they must not surface in the
-        // historical-admin view.
-        let log = AuditLog::in_memory(KeyPair::generate());
-        let session = SessionId::from_uuid(uuid::Uuid::nil());
-        log.append(
-            session.clone(),
-            AuditAction::McpToolCall {
-                server: "x".into(),
-                tool: "y".into(),
-                args_hash: astrid_crypto::ContentHash::from_bytes([0u8; 32]),
-            },
-            AuthorizationProof::System {
-                reason: "test".into(),
-            },
-            AuditOutcome::Success { details: None },
-        )
-        .await
-        .expect("append");
-        let entries = log.get_session_entries(&session).await.expect("read");
-        assert_eq!(entries.len(), 1);
-        assert!(
-            render_entry(&entries[0]).is_none(),
-            "McpToolCall must not render into the admin-history view"
-        );
-    }
-
-    #[tokio::test]
-    async fn render_admin_request_round_trips() {
-        let log = AuditLog::in_memory(KeyPair::generate());
-        let session = SessionId::from_uuid(uuid::Uuid::nil());
-        log.append_with_principal(
-            session.clone(),
-            PrincipalId::new("admin").unwrap(),
-            admin_action("AgentDelete", Some("alice")),
-            AuthorizationProof::System {
-                reason: "test".into(),
-            },
-            AuditOutcome::Success { details: None },
-        )
-        .await
-        .expect("append");
-        let entries = log.get_session_entries(&session).await.expect("read");
-        let view = render_entry(&entries[0]).expect("admin entry must render");
-        assert_eq!(view.method.as_deref(), Some("AgentDelete"));
-        assert_eq!(view.principal.as_deref(), Some("admin"));
-        assert_eq!(view.target_principal.as_deref(), Some("alice"));
-        assert_eq!(view.outcome, "success");
-    }
-
-    #[tokio::test]
-    async fn pagination_narrows_live_without_hiding_the_callers_records() {
-        let log = AuditLog::in_memory(KeyPair::generate());
-        let session = SessionId::from_uuid(uuid::Uuid::nil());
-        for (principal, method) in [
-            ("alice", "AliceOwnAfterNarrowing"),
-            ("bob", "BobHiddenAfterNarrowing"),
-            ("bob", "BobVisibleBeforeNarrowing"),
-        ] {
-            log.append_with_principal(
-                session.clone(),
-                PrincipalId::new(principal).unwrap(),
-                admin_action(method, None),
-                AuthorizationProof::System {
-                    reason: "test".into(),
-                },
-                AuditOutcome::Success { details: None },
-            )
-            .await
-            .expect("append");
-        }
-        let mut entries = log.get_session_entries(&session).await.expect("read");
-        entries.reverse();
-
-        let checks = Arc::new(AtomicUsize::new(0));
-        let checks_for_probe = Arc::clone(&checks);
-        let capability_probe = super::super::events::CapabilityProbe::new(move |_, _, _| {
-            checks_for_probe.fetch_add(1, Ordering::SeqCst) == 0
-        });
-        let caller = PrincipalId::new("alice").unwrap();
-        let access = AuditAccess {
-            capability_probe: &capability_probe,
-            caller_principal: &caller,
-            device_key_id: Some("0123456789abcdef"),
-            requested_principal: None,
-        };
-
-        let (page, _) = paginate_page(
-            entries,
-            &AuditQuery::default(),
-            &access,
-            (None, 0, None),
-            DEFAULT_LIMIT,
-        );
-        let methods: Vec<_> = page
-            .iter()
-            .filter_map(|entry| entry.method.as_deref())
-            .collect();
-
-        assert_eq!(
-            methods,
-            vec!["BobVisibleBeforeNarrowing", "AliceOwnAfterNarrowing"]
-        );
-        assert!(!methods.contains(&"BobHiddenAfterNarrowing"));
-    }
-
-    #[tokio::test]
-    async fn pagination_cursor_survives_live_narrowing_with_same_second_batch() {
-        let log = AuditLog::in_memory(KeyPair::generate());
-        let session = SessionId::from_uuid(uuid::Uuid::nil());
-        for (principal, method) in [
-            ("alice", "AliceOlder"),
-            ("alice", "AliceVisibleAfterNarrowing"),
-            ("bob", "BobHiddenAfterNarrowing"),
-            ("bob", "BobVisibleBeforeNarrowing"),
-        ] {
-            log.append_with_principal(
-                session.clone(),
-                PrincipalId::new(principal).unwrap(),
-                admin_action(method, None),
-                AuthorizationProof::System {
-                    reason: "test".into(),
-                },
-                AuditOutcome::Success { details: None },
-            )
-            .await
-            .expect("append");
-        }
-        let mut entries = log.get_session_entries(&session).await.expect("read");
-        entries.reverse();
-        let same_second = Timestamp::from_datetime(
-            chrono::Utc
-                .timestamp_opt(1_700_000_000, 0)
-                .single()
-                .unwrap(),
-        );
-        let next_second = Timestamp::from_datetime(
-            chrono::Utc
-                .timestamp_opt(1_699_999_999, 0)
-                .single()
-                .unwrap(),
-        );
-        for entry in &mut entries[..3] {
-            entry.timestamp = same_second;
-        }
-        entries[3].timestamp = next_second;
-
-        let firehose_probe = super::super::events::CapabilityProbe::new(|_, _, _| true);
-        let self_only_probe = super::super::events::CapabilityProbe::new(|_, _, _| false);
-        let caller = PrincipalId::new("alice").unwrap();
-        let firehose_access = AuditAccess {
-            capability_probe: &firehose_probe,
-            caller_principal: &caller,
-            device_key_id: Some("0123456789abcdef"),
-            requested_principal: None,
-        };
-        let self_only_access = AuditAccess {
-            capability_probe: &self_only_probe,
-            caller_principal: &caller,
-            device_key_id: Some("0123456789abcdef"),
-            requested_principal: None,
-        };
-
-        let (page_one, next_cursor) = paginate_page(
-            entries.clone(),
-            &AuditQuery::default(),
-            &firehose_access,
-            (None, 0, None),
-            1,
-        );
-        assert_eq!(
-            page_one[0].method.as_deref(),
-            Some("BobVisibleBeforeNarrowing")
-        );
-        assert_eq!(next_cursor.as_deref(), Some("1700000000_1_all"));
-
-        let (cursor_ts, cursor_offset, cursor_scope) =
-            parse_cursor(next_cursor.as_deref()).expect("page-one cursor parses");
-        validate_cursor_scope(cursor_scope.as_ref(), &self_only_access)
-            .expect("all-scope cursor may narrow to self-only");
-        let (page_two, next_cursor) = paginate_page(
-            entries.clone(),
-            &AuditQuery::default(),
-            &self_only_access,
-            (cursor_ts, cursor_offset, cursor_scope),
-            1,
-        );
-        assert_eq!(
-            page_two[0].method.as_deref(),
-            Some("AliceVisibleAfterNarrowing")
-        );
-        assert_eq!(next_cursor.as_deref(), Some("1700000000_3_p616c696365"));
-
-        let (cursor_ts, cursor_offset, cursor_scope) =
-            parse_cursor(next_cursor.as_deref()).expect("page-two cursor parses");
-        validate_cursor_scope(cursor_scope.as_ref(), &self_only_access)
-            .expect("self-only cursor continues under same scope");
-        let (page_three, next_cursor) = paginate_page(
-            entries.clone(),
-            &AuditQuery::default(),
-            &self_only_access,
-            (cursor_ts, cursor_offset, cursor_scope),
-            1,
-        );
-        assert_eq!(page_three[0].method.as_deref(), Some("AliceOlder"));
-        assert_eq!(next_cursor.as_deref(), Some("1699999999_1_p616c696365"));
-
-        let (cursor_ts, cursor_offset, cursor_scope) =
-            parse_cursor(next_cursor.as_deref()).expect("page-three cursor parses");
-        validate_cursor_scope(cursor_scope.as_ref(), &self_only_access)
-            .expect("self-only cursor continues under same scope");
-        let (page_four, next_cursor) = paginate_page(
-            entries,
-            &AuditQuery::default(),
-            &self_only_access,
-            (cursor_ts, cursor_offset, cursor_scope),
-            1,
-        );
-        assert!(page_four.is_empty());
-        assert_eq!(next_cursor.as_deref(), None);
-    }
-
-    #[test]
-    fn parse_cursor_handles_v1_v2_and_v3_shapes() {
-        // v1 (legacy): bare integer, no underscore — offset
-        // defaults to 0. We accept this shape so v0.7.0 cursors
-        // already in flight don't fail the next paginated fetch.
-        let (ts, off, scope) = parse_cursor(Some("1700000000")).expect("bare ts parses");
-        assert_eq!(ts, Some(1_700_000_000));
-        assert_eq!(off, 0);
-        assert_eq!(scope, None);
-
-        // v2: `<ts>_<offset>` — same-second batches resume cleanly
-        // without losing or duplicating entries across the page
-        // boundary.
-        let (ts, off, scope) = parse_cursor(Some("1700000000_3")).expect("v2 cursor parses");
-        assert_eq!(ts, Some(1_700_000_000));
-        assert_eq!(off, 3);
-        assert_eq!(scope, None);
-
-        // v3: `<ts>_<offset>_<scope>` — carries the last page's
-        // effective scope so incompatible widens fail closed.
-        let (ts, off, scope) =
-            parse_cursor(Some("1700000000_3_p616c696365")).expect("v3 cursor parses");
-        assert_eq!(ts, Some(1_700_000_000));
-        assert_eq!(off, 3);
-        assert_eq!(
-            scope,
-            Some(CursorScope::Principal(PrincipalId::new("alice").unwrap()))
-        );
-
-        // None: no cursor → no positioning, start from newest.
-        let (ts, off, scope) = parse_cursor(None).expect("None passes");
-        assert_eq!(ts, None);
-        assert_eq!(off, 0);
-        assert_eq!(scope, None);
-
-        // Garbage rejected with `BadRequest`.
-        assert!(parse_cursor(Some("not-a-number")).is_err());
-        assert!(parse_cursor(Some("123_not-a-number")).is_err());
-        assert!(parse_cursor(Some("not-a-number_4")).is_err());
-    }
-
-    #[tokio::test]
-    async fn pagination_cursor_rejects_scope_widening_after_self_only_page() {
-        let log = AuditLog::in_memory(KeyPair::generate());
-        let session = SessionId::from_uuid(uuid::Uuid::nil());
-        for (principal, method) in [
-            ("alice", "AliceOlder"),
-            ("alice", "AliceVisibleWhileScoped"),
-            ("bob", "BobNewlyVisibleAfterWidening"),
-        ] {
-            log.append_with_principal(
-                session.clone(),
-                PrincipalId::new(principal).unwrap(),
-                admin_action(method, None),
-                AuthorizationProof::System {
-                    reason: "test".into(),
-                },
-                AuditOutcome::Success { details: None },
-            )
-            .await
-            .expect("append");
-        }
-        let mut entries = log.get_session_entries(&session).await.expect("read");
-        entries.reverse();
-        let same_second = Timestamp::from_datetime(
-            chrono::Utc
-                .timestamp_opt(1_700_000_000, 0)
-                .single()
-                .unwrap(),
-        );
-        let next_second = Timestamp::from_datetime(
-            chrono::Utc
-                .timestamp_opt(1_699_999_999, 0)
-                .single()
-                .unwrap(),
-        );
-        for entry in &mut entries[..2] {
-            entry.timestamp = same_second;
-        }
-        entries[2].timestamp = next_second;
-
-        let self_only_probe = super::super::events::CapabilityProbe::new(|_, _, _| false);
-        let firehose_probe = super::super::events::CapabilityProbe::new(|_, _, _| true);
-        let caller = PrincipalId::new("alice").unwrap();
-        let self_only_access = AuditAccess {
-            capability_probe: &self_only_probe,
-            caller_principal: &caller,
-            device_key_id: Some("0123456789abcdef"),
-            requested_principal: None,
-        };
-        let firehose_access = AuditAccess {
-            capability_probe: &firehose_probe,
-            caller_principal: &caller,
-            device_key_id: Some("0123456789abcdef"),
-            requested_principal: None,
-        };
-
-        let (page_one, next_cursor) = paginate_page(
-            entries,
-            &AuditQuery::default(),
-            &self_only_access,
-            (None, 0, None),
-            1,
-        );
-        assert_eq!(
-            page_one[0].method.as_deref(),
-            Some("AliceVisibleWhileScoped")
-        );
-        let (_, _, cursor_scope) = parse_cursor(next_cursor.as_deref()).expect("cursor parses");
-        let err = validate_cursor_scope(cursor_scope.as_ref(), &firehose_access)
-            .expect_err("widening must fail closed");
-        assert!(
-            err.to_string()
-                .contains("restart pagination without a cursor"),
-            "unexpected error: {err}"
-        );
-    }
-
-    #[tokio::test]
-    async fn pagination_clips_mid_page_when_scope_rewidens() {
-        let log = AuditLog::in_memory(KeyPair::generate());
-        let session = SessionId::from_uuid(uuid::Uuid::nil());
-        for (principal, method) in [
-            ("bob", "BobVisibleAfterRewidening"),
-            ("alice", "AliceVisibleWhileScoped"),
-            ("bob", "BobHiddenWhileScoped"),
-            ("bob", "BobVisibleBeforeNarrowing"),
-        ] {
-            log.append_with_principal(
-                session.clone(),
-                PrincipalId::new(principal).unwrap(),
-                admin_action(method, None),
-                AuthorizationProof::System {
-                    reason: "test".into(),
-                },
-                AuditOutcome::Success { details: None },
-            )
-            .await
-            .expect("append");
-        }
-        let mut entries = log.get_session_entries(&session).await.expect("read");
-        entries.reverse();
-
-        let checks = Arc::new(AtomicUsize::new(0));
-        let checks_for_probe = Arc::clone(&checks);
-        let capability_probe = super::super::events::CapabilityProbe::new(move |_, _, _| {
-            matches!(checks_for_probe.fetch_add(1, Ordering::SeqCst), 0 | 3)
-        });
-        let caller = PrincipalId::new("alice").unwrap();
-        let access = AuditAccess {
-            capability_probe: &capability_probe,
-            caller_principal: &caller,
-            device_key_id: Some("0123456789abcdef"),
-            requested_principal: None,
-        };
-
-        let (page, next_cursor) =
-            paginate_page(entries, &AuditQuery::default(), &access, (None, 0, None), 3);
-        let methods: Vec<_> = page
-            .iter()
-            .filter_map(|entry| entry.method.as_deref())
-            .collect();
-
-        assert_eq!(
-            methods,
-            vec!["BobVisibleBeforeNarrowing", "AliceVisibleWhileScoped"]
-        );
-        let (cursor_ts, cursor_offset, cursor_scope) =
-            parse_cursor(next_cursor.as_deref()).expect("cursor parses");
-        assert!(
-            cursor_ts.is_some(),
-            "clipped page must still return a cursor"
-        );
-        assert_eq!(cursor_offset, 3);
-        assert_eq!(
-            cursor_scope,
-            Some(CursorScope::Principal(PrincipalId::new("alice").unwrap()))
-        );
-    }
-}
+mod tests;
